ct/gc: move GC interval related functions to ct/gc
This commit moves GC interfal related functions from the
ctmap package into the ctmap/gc package.

In addition, test-only functions are moved into the test files and
constants have been unexported and defined as consts (not vars).
